Add refresh token functionality and adjust JWT lifespan
The main update is the addition of refresh tokens to aid in user authentication, associated with each user in the ApplicationUser model. The JSON Web Token (JWT) lifespan was also shortened from 360 minutes to 30 minutes and refresh token expiry set to 10080 minutes (7 days). A new endpoint was added to the Api Gateway and necessary migrations were created and applied.